# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: symbols
    description: Symbol universe management
  - name: candles
    description: OHLCV candle data
  - name: metrics
    description: Market metrics
  - name: scanners
    description: Scanner management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /symbols:
    get:
      operationId: listSymbols
      tags: [symbols]
      summary: List active symbols
      description: Returns all active symbols with score, iv_rank, and last_scored
      responses:
        "200":
          description: Active symbol list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SymbolRecord"
  /symbols/{symbol}:
    post:
      operationId: addSymbol
      tags: [symbols]
      summary: Add symbol to universe
      description: Adds a symbol, opens DXLink quote subscription and caches metrics
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Symbol added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SymbolRecord"
        "409":
          description: Symbol already active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeSymbol
      tags: [symbols]
      summary: Remove symbol from universe
      description: Removes a symbol and closes its DXLink subscription
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Symbol removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Symbol not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /candles/{symbol}/{timeframe}:
    get:
      operationId: getCandles
      tags: [candles]
      summary: Get OHLCV candles
      description: Returns OHLCV array for the given symbol and timeframe (1d, 60m, 15m)
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: timeframe
          in: path
          required: true
          schema:
            type: string
            enum: [1d, 60m, 15m]
      responses:
        "200":
          description: Candle data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"
        "400":
          description: Invalid timeframe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /metrics/{symbol}:
    get:
      operationId: getMetrics
      tags: [metrics]
      summary: Get market metrics
      description: Returns IV rank, IVx, earnings date, and lendability (5-minute TTL cache)
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Market metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketMetrics"
  /scanners:
    get:
      operationId: listScanners
      tags: [scanners]
      summary: List all scanners
      description: Returns all scanners with their status, last run time, and result count
      responses:
        "200":
          description: Scanner list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScannerSummary"
    post:
      operationId: createScanner
      tags: [scanners]
      summary: Create scanner
      description: Creates a new scanner with filter rules and a run interval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScannerRequest"
      responses:
        "201":
          description: Scanner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScannerDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scanners/{id}:
    get:
      operationId: getScanner
      tags: [scanners]
      summary: Get scanner
      description: Returns full scanner config including filter rules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scanner detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScannerDetail"
        "404":
          description: Scanner not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateScanner
      tags: [scanners]
      summary: Update scanner
      description: Updates an existing scanner's config; reschedules if needed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScannerRequest"
      responses:
        "200":
          description: Scanner updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScannerDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Scanner not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteScanner
      tags: [scanners]
      summary: Delete scanner
      description: Deletes a scanner and removes its schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scanner deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Scanner not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scanners/{id}/results:
    get:
      operationId: getScannerResults
      tags: [scanners]
      summary: Get scanner results
      description: Returns the current snapshot of passing symbols with their metrics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scanner results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScannerResultItem"
        "404":
          description: Scanner not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        symbolCount:
          type: integer
          description: Number of active symbols tracked in the DB
        connectedSymbolCount:
          type: integer
          description: Number of symbols with live DXLink quote subscriptions
      required:
        - status
    SymbolRecord:
      type: object
      properties:
        symbol:
          type: string
        addedAt:
          type: integer
        lastScored:
          type: integer
          nullable: true
        ivRank:
          type: number
          nullable: true
        ivx:
          type: number
          nullable: true
        earningsDate:
          type: string
          nullable: true
        score:
          type: number
          nullable: true
        active:
          type: integer
      required:
        - symbol
        - addedAt
        - active
    Candle:
      type: object
      properties:
        time:
          type: number
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
      required:
        - time
        - open
        - high
        - low
        - close
        - volume
    MarketMetrics:
      type: object
      properties:
        symbol:
          type: string
        ivRank:
          type: number
          nullable: true
        ivx:
          type: number
          nullable: true
        earningsDate:
          type: string
          nullable: true
        lendability:
          type: string
          nullable: true
      required:
        - symbol
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    FilterRule:
      oneOf:
        - description: Filter rule for numeric metrics
          type: object
          properties:
            metric:
              type: string
              enum: [ivRank, ivx, score, earningsDate, price, volume, open, high, low, ivPercentile, liquidityRank, beta, marketCap, callVolume, putVolume, callOpenInterest, putOpenInterest]
            operator:
              type: string
              enum: [">", ">=", "<", "<=", "=", "!=", "between"]
            value:
              type: number
            value2:
              type: number
              nullable: true
              description: Required when operator is "between"
          required:
            - metric
            - operator
            - value
        - description: Filter rule for text metrics (lendability, instrumentType)
          type: object
          properties:
            metric:
              type: string
              enum: [lendability, instrumentType]
            operator:
              type: string
              enum: ["=", "!="]
            value:
              type: string
          required:
            - metric
            - operator
            - value
    ScannerSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        intervalSeconds:
          type: integer
        enabled:
          type: boolean
        lastRunAt:
          type: integer
          nullable: true
        resultCount:
          type: integer
        filterCount:
          type: integer
        createdAt:
          type: integer
      required:
        - id
        - name
        - intervalSeconds
        - enabled
        - resultCount
        - filterCount
        - createdAt
    ScannerDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        filters:
          type: array
          items:
            $ref: "#/components/schemas/FilterRule"
        intervalSeconds:
          type: integer
        enabled:
          type: boolean
        lastRunAt:
          type: integer
          nullable: true
        createdAt:
          type: integer
      required:
        - id
        - name
        - filters
        - intervalSeconds
        - enabled
        - createdAt
    CreateScannerRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        filters:
          type: array
          items:
            $ref: "#/components/schemas/FilterRule"
        intervalSeconds:
          type: integer
          default: 300
        enabled:
          type: boolean
          default: true
      required:
        - name
        - filters
    UpdateScannerRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        filters:
          type: array
          items:
            $ref: "#/components/schemas/FilterRule"
        intervalSeconds:
          type: integer
        enabled:
          type: boolean
    ScannerResultItem:
      type: object
      properties:
        symbol:
          type: string
        ivRank:
          type: number
          nullable: true
        ivx:
          type: number
          nullable: true
        score:
          type: number
          nullable: true
        lendability:
          type: string
          nullable: true
        earningsDate:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        volume:
          type: number
          nullable: true
        open:
          type: number
          nullable: true
        high:
          type: number
          nullable: true
        low:
          type: number
          nullable: true
        ivPercentile:
          type: number
          nullable: true
        liquidityRank:
          type: number
          nullable: true
        beta:
          type: number
          nullable: true
        marketCap:
          type: number
          nullable: true
        callVolume:
          type: number
          nullable: true
        putVolume:
          type: number
          nullable: true
        callOpenInterest:
          type: number
          nullable: true
        putOpenInterest:
          type: number
          nullable: true
        instrumentType:
          type: string
          nullable: true
        updatedAt:
          type: integer
      required:
        - symbol
        - updatedAt