# lib/api-zod/src/generated/types/createScannerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FilterRule } from "./filterRule";

export interface CreateScannerRequest {
  name: string;
  description?: string | null;
  filters: FilterRule[];
  intervalSeconds?: number;
  enabled?: boolean;
}