# lib/api-zod/src/generated/types/filterRule.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type FilterRule =
  | {
      metric:
        | "ivRank"
        | "ivx"
        | "score"
        | "earningsDate"
        | "price"
        | "volume"
        | "open"
        | "high"
        | "low";
      operator: ">" | ">=" | "<" | "<=" | "=" | "!=" | "between";
      value: number;
      /** Required when operator is "between" */
      value2?: number | null;
    }
  | {
      metric: "lendability";
      operator: "=" | "!=";
      value: string;
    };