# lib/api-zod/src/generated/types/scannerDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FilterRule } from "./filterRule";

export interface ScannerDetail {
  id: string;
  name: string;
  description?: string | null;
  filters: FilterRule[];
  intervalSeconds: number;
  enabled: boolean;
  lastRunAt?: number | null;
  createdAt: number;
}