# lib/api-zod/src/generated/types/scannerResultItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScannerResultItem {
  symbol: string;
  ivRank?: number | null;
  ivx?: number | null;
  score?: number | null;
  lendability?: string | null;
  earningsDate?: string | null;
  price?: number | null;
  volume?: number | null;
  open?: number | null;
  high?: number | null;
  low?: number | null;
  updatedAt: number;
}